changefeedccl: move PTS code to dedicated files
This change moves tests and helpers related to managing PTS
records to their own dedicated files. This change also
adds comments to each test related to PTS records.

Epic: None
Release note: None
