+++ pkg/ccl/changefeedccl/protected_timestamps.go
@@ -1,0 +1,76 @@
+// Copyright 2023 The Cockroach Authors.
+//
+// Licensed as a CockroachDB Enterprise file under the Cockroach Community
+// License (the "License"); you may not use this file except in compliance with
+// the License. You may obtain a copy of the License at
+//
+//     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
+
+package changefeedccl
+
+import (
+	"context"
+
+	"github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl/changefeedbase"
+	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
+	"github.com/cockroachdb/cockroach/pkg/jobs/jobsprotectedts"
+	"github.com/cockroachdb/cockroach/pkg/keys"
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/protectedts/ptpb"
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
+	"github.com/cockroachdb/cockroach/pkg/util/hlc"
+	"github.com/cockroachdb/cockroach/pkg/util/log"
+	"github.com/cockroachdb/cockroach/pkg/util/uuid"
+)
+
+// createProtectedTimestampRecord will create a record to protect the spans for
+// this changefeed at the resolved timestamp.
+func createProtectedTimestampRecord(
+	ctx context.Context,
+	codec keys.SQLCodec,
+	jobID jobspb.JobID,
+	targets changefeedbase.Targets,
+	resolved hlc.Timestamp,
+) *ptpb.Record {
+	ptsID := uuid.MakeV4()
+	deprecatedSpansToProtect := makeSpansToProtect(codec, targets)
+	targetToProtect := makeTargetToProtect(targets)
+
+	log.VEventf(ctx, 2, "creating protected timestamp %v at %v", ptsID, resolved)
+	return jobsprotectedts.MakeRecord(
+		ptsID, int64(jobID), resolved, deprecatedSpansToProtect,
+		jobsprotectedts.Jobs, targetToProtect)
+}
+
+func makeTargetToProtect(targets changefeedbase.Targets) *ptpb.Target {
+	// NB: We add 1 because we're also going to protect system.descriptors.
+	// We protect system.descriptors because a changefeed needs all of the history
+	// of table descriptors to version data.
+	tablesToProtect := make(descpb.IDs, 0, targets.NumUniqueTables()+1)
+	_ = targets.EachTableID(func(id descpb.ID) error {
+		tablesToProtect = append(tablesToProtect, id)
+		return nil
+	})
+	tablesToProtect = append(tablesToProtect, keys.DescriptorTableID)
+	return ptpb.MakeSchemaObjectsTarget(tablesToProtect)
+}
+
+func makeSpansToProtect(codec keys.SQLCodec, targets changefeedbase.Targets) []roachpb.Span {
+	// NB: We add 1 because we're also going to protect system.descriptors.
+	// We protect system.descriptors because a changefeed needs all of the history
+	// of table descriptors to version data.
+	spansToProtect := make([]roachpb.Span, 0, targets.NumUniqueTables()+1)
+	addTablePrefix := func(id uint32) {
+		tablePrefix := codec.TablePrefix(id)
+		spansToProtect = append(spansToProtect, roachpb.Span{
+			Key:    tablePrefix,
+			EndKey: tablePrefix.PrefixEnd(),
+		})
+	}
+	_ = targets.EachTableID(func(id descpb.ID) error {
+		addTablePrefix(uint32(id))
+		return nil
+	})
+	addTablePrefix(keys.DescriptorTableID)
+	return spansToProtect
+}
